+++ test/unit/deref.test.js
@@ -100,6 +100,14 @@ describe('DEREF FUNCTION TESTS ', function() {
                   name: {
                     type: 'string'
                   }
+                },
+                additionalProperties: {
+                  type: 'object',
+                  properties: {
+                    hello: {
+                      type: 'string'
+                    }
+                  }
                 }
               }
             }
@@ -119,17 +127,18 @@ describe('DEREF FUNCTION TESTS ', function() {
           _.cloneDeep(componentsAndPaths), {}, 'VALIDATION'),
         output_emptyObject = deref.resolveRefs(schemaWithEmptyObject, parameterSource, _.cloneDeep(componentsAndPaths)),
         output_additionalProps = deref.resolveRefs(schemaWithAdditionPropRef, parameterSource,
-          _.cloneDeep(componentsAndPaths), {}, 'VALIDATION');
+          _.cloneDeep(componentsAndPaths), {}, 'VALIDATION'),
+        output_additionalPropsOverride;
 
       expect(output).to.deep.include({ type: 'object',
         required: ['id'],
-        properties: { id: { default: '<long>', type: 'integer' } } });
+        properties: { id: { default: '<long>', format: 'int64', type: 'integer' } } });
 
       expect(output_validation).to.deep.include({ anyOf: [
         { type: 'object',
           required: ['id'],
           description: 'Schema 2',
-          properties: { id: { type: 'integer' } }
+          properties: { id: { format: 'int64', type: 'integer' } }
         }, {
           type: 'object',
           properties: { emailField: { type: 'string', format: 'email' } }
@@ -138,7 +147,7 @@ describe('DEREF FUNCTION TESTS ', function() {
 
       expect(output_withdot).to.deep.include({ type: 'object',
         required: ['id'],
-        properties: { id: { default: '<long>', type: 'integer' } } });
+        properties: { id: { default: '<long>', format: 'int64', type: 'integer' } } });
 
       expect(output_customFormat).to.deep.include({ type: 'object',
         properties: { emailField: { default: '<email>', format: 'email', type: 'string' } } });
@@ -147,7 +156,7 @@ describe('DEREF FUNCTION TESTS ', function() {
         type: 'object',
         description: 'Schema 2',
         properties: {
-          id: { default: '<long>', type: 'integer' },
+          id: { default: '<long>', format: 'int64', type: 'integer' },
           test_prop: { default: '<string>', type: 'string' }
         }
       });
@@ -170,42 +179,15 @@ describe('DEREF FUNCTION TESTS ', function() {
       // additionalProperties $ref should be resolved
       expect(output_additionalProps).to.deep.include(componentsAndPaths.components.schemas.schemaAdditionalProps);
 
-      done();
-    });
+      // add default to above resolved schema
+      output_additionalProps.additionalProperties.properties.hello.default = '<string>';
 
-    it('should populate schemaResolutionCache having key as the ref provided', function (done) {
-      var schema = {
-          $ref: '#/components/schema/request'
-        },
-        componentsAndPaths = {
-          components: {
-            schema: {
-              request: {
-                properties: {
-                  name: {
-                    type: 'string',
-                    example: 'example name'
-                  }
-                }
-              }
-            }
-          },
-          concreteUtils: schemaUtils30X
-        },
-        parameterSource = 'REQUEST',
-        schemaResolutionCache = {},
-        resolvedSchema = deref.resolveRefs(schema, parameterSource, componentsAndPaths, schemaResolutionCache);
-      expect(_.get(schemaResolutionCache, ['#/components/schema/request', 'schema'])).to.deep.equal(resolvedSchema);
-      expect(resolvedSchema).to.deep.equal({
-        type: 'object',
-        properties: {
-          name: {
-            type: 'string',
-            example: 'example name',
-            default: '<string>'
-          }
-        }
-      });
+      output_additionalPropsOverride = deref.resolveRefs(schemaWithAdditionPropRef, parameterSource,
+        _.cloneDeep(componentsAndPaths), {}, 'VALIDATION');
+
+      // override should not affect newly resolved schema
+      expect(output_additionalPropsOverride).to.deep.include(
+        componentsAndPaths.components.schemas.schemaAdditionalProps);
       done();
     });
 
@@ -219,9 +201,6 @@ describe('DEREF FUNCTION TESTS ', function() {
         ],
         nonSupportedFormats = [
           { type: 'integer', format: 'int32' },
-          { type: 'integer', format: 'int64' },
-          { type: 'number', format: 'float' },
-          { type: 'number', format: 'double' },
           { type: 'string', format: 'byte' },
           { type: 'string', format: 'binary' },
           { type: 'string', format: 'password' },
@@ -274,87 +253,63 @@ describe('DEREF FUNCTION TESTS ', function() {
       done();
     });
 
-    it('should correctly resolve schema from schemaResoltionCache based on schema resolution level', function (done) {
-      let schema = {
-          $ref: '#/components/schemas/schemaUsed'
-        },
-        consumerSchema = {
+    it('should not contain readOnly properties in resolved schema if they are not contained' +
+      ' in resolved schema', function(done) {
+      var schema = {
           type: 'object',
-          properties: { level2: {
-            type: 'object',
-            properties: { level3: {
-              type: 'object',
-              properties: { level4: {
-                type: 'object',
-                properties: { level5: {
-                  type: 'object',
-                  properties: { level6: {
-                    type: 'object',
-                    properties: { level7: {
-                      type: 'object',
-                      properties: { level8: {
-                        type: 'object',
-                        properties: { level9: { $ref: '#/components/schemas/schemaUsed' } }
-                      } }
-                    } }
-                  } }
-                } }
-              } }
-            } }
-          } }
-        },
-        componentsAndPaths = {
-          components: {
-            schemas: {
-              schemaUsed: {
-                'type': 'object',
-                'required': [
-                  'id',
-                  'name'
-                ],
-                'properties': {
-                  'id': {
-                    'type': 'integer',
-                    'format': 'int64'
-                  },
-                  'name': {
-                    'type': 'string'
-                  },
-                  'tag': {
-                    'type': 'string'
-                  }
-                }
-              }
+          required: ['id', 'name'],
+          properties: {
+            id: {
+              type: 'integer',
+              format: 'int64',
+              readOnly: true
+            },
+            name: {
+              type: 'string'
+            },
+            tag: {
+              type: 'string',
+              writeOnly: true
             }
-          },
-          concreteUtils: schemaUtils30X
+          }
         },
         parameterSource = 'REQUEST',
-        schemaResoltionCache = {},
-        resolvedConsumerSchema,
-        resolvedSchema;
+        output;
 
-      resolvedConsumerSchema = deref.resolveRefs(consumerSchema, parameterSource, componentsAndPaths,
-        schemaResoltionCache);
+      output = deref.resolveRefs(schema, parameterSource, { concreteUtils: schemaUtils30X });
+      expect(output.type).to.equal('object');
+      expect(output.properties).to.not.haveOwnProperty('id');
+      expect(output.required).to.not.include('id');
+      done();
+    });
 
-      // Consumer schema contains schema at nesting level 9, which results in impartial resolution of schema
-      expect(_.get(schemaResoltionCache, ['#/components/schemas/schemaUsed', 'resLevel'])).to.eql(9);
-      expect(_.get(resolvedConsumerSchema, _.join(_.map(_.range(1, 10), (ele) => {
-        return `properties.level${ele}`;
-      }), '.'))).to.not.deep.equal(componentsAndPaths.components.schemas.schemaUsed);
-      expect(_.get(schemaResoltionCache, ['#/components/schemas/schemaUsed', 'schema'])).to.not.deep
-        .equal(componentsAndPaths.components.schemas.schemaUsed);
-      resolvedSchema = deref.resolveRefs(schema, parameterSource, componentsAndPaths, schemaResoltionCache);
-      // Restoring the original format as it is deleted if not supported by json-schema-faker and ajv
-      resolvedSchema.properties.id.format = 'int64';
+    it('should not contain writeOnly properties in resolved schema if they are not contained' +
+      ' in resolved schema', function(done) {
+      var schema = {
+          type: 'object',
+          required: ['id', 'tag'],
+          properties: {
+            id: {
+              type: 'integer',
+              format: 'int64',
+              readOnly: true
+            },
+            name: {
+              type: 'string'
+            },
+            tag: {
+              type: 'string',
+              writeOnly: true
+            }
+          }
+        },
+        parameterSource = 'RESPONSE',
+        output;
 
-      /**
-       * Even though schema cache contains schemaUsed as impartially cached,resolution were it's used again will
-       * depend on ongoing resolution level and schema is cached again if it's updated.
-       */
-      expect(resolvedSchema).to.deep.equal(componentsAndPaths.components.schemas.schemaUsed);
-      expect(_.get(schemaResoltionCache, ['#/components/schemas/schemaUsed', 'schema'])).to.deep
-        .equal(componentsAndPaths.components.schemas.schemaUsed);
+      output = deref.resolveRefs(schema, parameterSource, { concreteUtils: schemaUtils30X });
+      expect(output.type).to.equal('object');
+      expect(output.properties).to.not.haveOwnProperty('tag');
+      expect(output.required).to.not.include('tag');
       done();
     });
   });