+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # OpenAPI-Postman Changelog
 
+#### v3.2.0 (May 02, 2022)
+* Fixed some of critical and high level severity vulnerabilities.
+* Fixed issue [#10752](https://github.com/postmanlabs/postman-app-support/issues/10752) where deepObject style parameters were not generated correctly.
+* Fixed issue [#485](https://github.com/postmanlabs/openapi-to-postman/issues/485) where validateTransaction() returns result where mismatch path for transaction are incorrect.
+* Fixed issue [#485](https://github.com/postmanlabs/openapi-to-postman/issues/485) where validateTransaction() returns result where mismatch path for transaction are incorrect.
+
+#### v3.1.0 (March 04, 2022)
+* Removed usage of schema resolution cache to avoid incorrect resolution.
+* Fixed issue where newly converted collection had mismatches from validateTransaction() API.
+* Fixed issue where falsy example were not getting converted correctly in generated collection.
+* Fixed issue where readOnly and writeOnly properties were not removed from required in resolved schema.
+* Fixed issue where if schema had pattern as property, conversion failed with error.
+
 #### v3.0.0 (Feb 11, 2022)
 * Add support for OpenAPI 3.1.x https://www.openapis.org/blog/2021/02/18/openapi-specification-3-1-released.
 * Separated schemUtils into common, 3.1, and 3.0 specific files.